Renames mncatdiscovery to primoui with BC alias, fixes #20
